fix(*): Fixing one more time the tests from CircleCI
Since these are old distros and releases, most of them are no longer
supported. Ubuntu version updated, and removed the old/unused versions

Using OpenJDK instead of Oracle Java from Ansiblebit now for the
CircleCI testing, it's necessary since most of the tests failures
happens in the Oracle Java installation, which is out of scope for the
continuous integration of this role.