refact(*): Fixing the CircleCI test suite
Including a wider variety of distros on the test suite.
Since the OpenJDK is being used as the base for tests,
it's easier to find it available on distros' packages indexes.

Therefore, gives more flexibility during the execution of tests on CI to
pick whatever distro comes to mind.

To make it even easier, the OpenJDK was included on the Docker images,
installed through automated and dynamic process on
[macunha1/docker-ansible](github.com/macunha1/docker-ansible/)