refact(*): Updating test suite to support latest Ansible version
CircleCI tests are now being executed against Ansible 2.9.6 and both
Python 2 and 3. This should facilitate the evolution of this role on the
next increments.

Installation of SE Linux Python bindings moved to Pip, using a
block/rescue to try/catch errors and try to recover installing from the
packages (as long as the indexes/package cache are updated on the OS).