chore(code lint): Updating rules for Ansible Lint
Supresses errors that are currently happening at Ansible Lint due to
changes at the default values for YAML Lint. As YAML Lint is executed
alone with a custom configuration file, it is safe to skip Ansible Lint
for YAML altogether.

Renaming the role as the naming convention changed from dashes to
underscores and since the last import to Ansible Galaxy it was enforced
to confluent_kafka (even though I don't agree nor see any benefit with
this change)